refactor: move session_manager to brokers/robinhood_session (#11)
Eliminates the reverse `brokers/ → tools/` architectural dependency by
relocating `tools/session_manager.py` to `brokers/robinhood_session.py`.
Updates all production imports (6 files) and test mock patch paths (3 files).
Adds a guard test verifying the new module is importable and the old path is gone.
